reset to empty proj with updated dependencies
