--- .npmignore
@@ -1,6 +1,0 @@
-tsconfig.json
-src
-test
-benchmark
-.*
-
