+++ old-example/old/intersections.ts
