+++ old-example/old/old-example/index.ts
