+++ old-src/graph-state.ts
