+++ src/index.ts
@@ -1,2 +1,1 @@
-export * from "./graph-view";
-export * from "./utils";
+export default null;
