+++ test/index.test.ts
@@ -1,0 +1,7 @@
+import { describe, it, expect } from "vitest";
+
+describe("blah", () => {
+  it("works", () => {
+    expect(5).toBeGreaterThan(0);
+  });
+});
